requirements.txt
torch==2.5.1
torchvision==0.20.1
model2vec==0.3.3
Pillow==11.0.0
torchmetrics
wandb